Mongolia

Tools for working with Mongodb data in Julia

Mongolia provides a Tables implementation for [Mongoc.Cursor].

Mongo data => Mongolian Tables

Mongolian query syntax

Mongolian Tables => Mongo data

SSH Tunnel

Misc. Tips

Memoize Expensive Queries

Index

Mongolia.NOISE_FIELDS

Mongolia.Database

Base.getproperty

Mongolia.bson

Mongolia.flatten_dicts

Mongolia.lookup

Mongolia.mfind

Mongolia.tomany

Mongolia.toparent

Mongolia.unflatten

Mongolia.NOISE_FIELDS — Constant

Mongo field names to be moved to the right via select!.

Mongolia.Database — Type

Wrap a Mongoc.Database and a Mongoc.Client.

Allows collections to be accessed as properties.

Base.getproperty — Method

Overload property access as shortcut for collections.

db.project == db["project"]

Mongolia.bson — Method

bson(pair...) => Mongoc.BSON
bson(AbstractVector{pair}) => Vector{Mongoc.BSON}

Construct a BSON object using keyword arguments or pairs to reduce quote clutter.

Examples

julia> bson(:metadata!pt => r"^mei.*1$", :project => +:in => ["proj1", "proj2"])
Mongoc.BSON with 2 entries:
  "metadata.pt" => Dict{Any, Any}("\$regex"=>"^mei.*1\$")
  "project"     => Dict{Any, Any}("\$in"=>Any["proj1", "proj2"])

julia> bson([
         +:match => :type=>"SUBMIT", 
         +:group => (:_id=>+:pt, :N=>+:sum=>1)
       ])
Mongoc.BSON with 2 entries:
  "0" => Dict{Any, Any}("\$match"=>Dict{Any, Any}("type"=>"SUBMIT"))
  "1" => Dict{Any, Any}("\$group"=>Dict{Any, Any}("_id"=>"\$pt", "N"=>Dict{Any,…
julia> bson(a=1, b=2)
Mongoc.BSON with 2 entries:
  "a" => 1
  "b" => 2

Mongolia.flatten_dicts — Method

flatten_dicts(cursor ; [separator = "!"])

Convert an iterable of nested Dict{String,Any} (such as a Cursor returned by Mongoc.find) into a dictonary of equal length columns.

The returned dictionary can be converted to a DataFrame.

Each field value that is a dictionary is replaced a field for every entry. The field names are the path.

Using "!" does not require quoting in Symbol names, so you can type :metadata!pt instead of "metadata.pt".

Option Arguments

separator : path separator for flattened column names.

Mongolia.lookup — Method

Construct lookup and optional unwind stages for a pipeline.

The defaults work for a to-many relation where the child collection's foreign key is the parent collection name.

Mongolia.mfind — Method

mfind(collection, queryitem...; limit=0, option...)
mfind(database, project_code, dataset_code, queritem...; limit=0, option...)

Query mongodb with a powerful short-hand syntax and return a dataframe.

Each queryitem can be a Pair or vector of pairs using bson syntax.

The vectors are flattened so you don't need ... after tomany[@ref].

If any item looks like a pipeline stage (starts with '$(Expr(:incomplete, "incomplete: invalid character literal"))

Examples

julia> db = database("mmendel",QA);

julia> mfind(db.project, :code=>+:regex=>"^test"; :limit=>1, :skip=>1)
1×12 DataFrame
 Row │ active  code    dateCreated              isDeleted  lastUpdated         ⋯
     │ Bool    String  DateTime                 Bool       DateTime            ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │   true  test2   2015-01-12T17:28:50.481      false  2015-01-12T17:28:50 ⋯
                                                               8 columns omitted
julia> configs = mfind(db.project, 
         :code=>+:regex=>"^test",
         tomany("project", "instrument", "instrumentConfig"; skipmissing=true)
         ; limit=3);

julia> configs[!, [:name, :instrumentConfig!name]]
3×2 DataFrame
 Row │ name            instrumentConfig!name 
     │ String          String                
─────┼───────────────────────────────────────
   1 │ Test Run        Protocol 1 Config
   2 │ Test James 102  test config #asdf
   3 │ Test Import     Branching Example

Mongolia.tomany — Method

tomany(parent, children...)

Construct a pipeline traversing a chain of to-many relations.

Examples

julia> tomany(:device, :devconfig, :config=>:trigger)

[...]

Mongolia.toparent — Method

toparent(key...; skipmissing=false)

Join with a chain of parents specified by the keys. o

Examples

mfind(db.participant, toparent(:project))
mfind(db.trigger. topparent(:configuration => :instrumentConfig))

Mongolia.unflatten — Method

unflatten(row)
unflatten(::Vector{row})
unflatten(::AbstractDataFrame)

Convert flattened mongo docs to its original shape