Mongolia

Tools for working with Mongodb data in Julia

Mongolia provides a Tables implementation for [Mongoc.Cursor].

Mongo data => Mongolian Tables

Mongolian query syntax

Mongolian Tables => Mongo data

SSH Tunnel

Misc. Tips

Memoize Expensive Queries

Index

Mongolia.bson

Mongolia.tomany

Mongolia.toparent

Mongolia.unflatten

Mongolia.bson — Method

bson(pair...) => Mongoc.BSON
bson(AbstractVector{pair}) => Vector{Mongoc.BSON}

Construct a BSON object using keyword arguments or pairs to reduce quote clutter.

Examples

julia> bson("metadata.pt" => "xyz", :project => +:in => ["proj1", "proj2"])
Mongoc.BSON with 2 entries:
  "metadata.pt" => "xyz"
  "project"     => Dict{Any, Any}("\$in"=>Any["proj1", "proj2"])

julia> bson([
         +:match => :type=>"SUBMIT", 
         +:group => (:_id=>+:pt, :N=>+:sum=>1)
       ])
Mongoc.BSON with 2 entries:
  "0" => Dict{Any, Any}("\$match"=>Dict{Any, Any}("type"=>"SUBMIT"))
  "1" => Dict{Any, Any}("\$group"=>Dict{Any, Any}("_id"=>"\$pt", "N"=>Dict{Any,…
julia> bson(a=1, b=2)
Mongoc.BSON with 2 entries:
  "a" => 1
  "b" => 2

Mongolia.tomany — Method

tomany(parent, children...)

Pipeline helper TODO

Examples

Mongolia.toparent — Method

toparent(key...; skipmissing=false)

Join with a chain of parents specified by the keys. o

Examples

pilrfind(db.participant, toparent(:project))
pilrfind(db.trigger. topparent(:configuration => :instrumentConfig))

Mongolia.unflatten — Method

Convert flattened mongo doc to original shape